requirements.txt
alembic==1.17.0
apturl==0.5.2
bcrypt==3.2.0
blinker==1.9.0
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
chartjs==1.2
click==8.3.0
colorama==0.4.4
command-not-found==0.3
contourpy==1.3.2
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
fasteners==0.14.1
Flask==2.2.5
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
fonttools==4.60.1
future==0.18.2
greenlet==3.2.4
gunicorn==21.2.0
gyp==0.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.6
joblib==1.5.2
keyring==23.5.0
kiwisolver==1.4.9
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==3.0.3
matplotlib==3.10.7
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.0
olefile==0.46
packaging==25.0
pandas==2.2.2
paramiko==2.9.3
pexpect==4.8.0
Pillow==10.0.0
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
pygame==2.6.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==3.2.5
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==4.0.9
requests==2.25.1
scikit-learn==1.3.2
scipy==1.15.3
SecretStorage==3.3.1
six==1.16.0
SQLAlchemy==2.0.43
systemd-python==234
threadpoolctl==3.6.0
tomli==2.3.0
typing_extensions==4.15.0
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
Werkzeug==2.3.6
WTForms==3.2.1
xdg==5
xkit==0.0.0
zipp==1.0.0
